Move some GraphQL requests to seperated files
