+++ sensemap/src/types/graphql/client.ts
